Allow nesting of single and double quotes and long comments
Previously the regular expression used in `split_quoted` would
erroneously consider a single and double quote as a pair.  This made it
impossible to have a comment with a nested quote.

Additionally for versions of iptables that do not support the `iptables
-C` check feature, long comments would not match, and single quoted
comments had to be tested using double quotes.  This is because
internally long comments are truncated to 255 characters and the
output of `iptables-save` and `-S` rebuilds the rule using double quotes
instead of single quotes for arguments that contain spaces.

This fixes the issue of nested quotes by splitting on  matching pairs of
single and double quotes.  This fixes the exists method for old versions
of iptables that do not support `-C` by performing a similar process of
first splitting the rule into its command parts, truncating long
strings, escaping and quoting arguments, and reassembling the rule.